herbie_vision/model/model_serving/requirements.txt
opencv-python
numpy
requests
flask
pillow
torch
torchvision
google-cloud-storage
herbie-vision